backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.1
PyJWT==2.10.1
sqlparse==0.5.3
whitenoise==6.7.0
gunicorn==23.0.0
dj-database-url==2.2.0
djangorestframework_simplejwt==5.5.1
